added separate file for thread_local_data
