+++ container_service_extension/__init__.py
@@ -4,6 +4,8 @@
 
 from container_service_extension.logger import configure_all_file_loggers
 from container_service_extension.logger import configure_null_logger
+from container_service_extension.thread_local_data import init_thread_local_data  # noqa: E501
 
+init_thread_local_data()
 configure_all_file_loggers()
 configure_null_logger()
